+++ airflow/contrib/kubernetes/pod_launcher.py
@@ -142,7 +142,7 @@ class PodLauncher(LoggingMixin):
         wait=tenacity.wait_exponential(),
         reraise=True
     )
-    def read_pod_logs(self, pod):
+    def read_pod_logs(self, pod, tail_lines=10):
 
         try:
             return self._client.read_namespaced_pod_log(
@@ -150,7 +150,7 @@ class PodLauncher(LoggingMixin):
                 namespace=pod.namespace,
                 container='base',
                 follow=True,
-                tail_lines=10,
+                tail_lines=tail_lines,
                 _preload_content=False
             )
         except BaseHTTPError as e:
