+++ airflow/version.py
@@ -18,5 +18,5 @@
 # under the License.
 #
 
-version = '1.10.4+twtr9.gke'
+version = '1.10.4+twtr9.gke6'
